modeling/requirements.txt
easyocr
transformers
accelerate
Pillow
opencv-python